# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 0.1.0
  description: Job Listing Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/seeker:
    get:
      operationId: getSeekerProfile
      tags: [profile]
      summary: Get the current seeker's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeekerProfile"
    put:
      operationId: updateSeekerProfile
      tags: [profile]
      summary: Update the current seeker's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeekerProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeekerProfile"
  /profile/employer:
    get:
      operationId: getEmployerProfile
      tags: [profile]
      summary: Get the current employer's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerProfile"
    put:
      operationId: updateEmployerProfile
      tags: [profile]
      summary: Update the current employer's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployerProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerProfile"
...
components:
  schemas:
...
    SeekerProfile:
      type: object
      required: [userId, fullName, headline, location, phone, bio, resumeUrl, skills, updatedAt]
      properties:
        userId: { type: string }
        fullName: { type: string }
        headline: { type: string }
        location: { type: string }
        phone: { type: string }
        bio: { type: string }
        resumeUrl: { type: string }
        skills:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
    SeekerProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        headline: { type: string }
        location: { type: string }
        phone: { type: string }
        bio: { type: string }
        resumeUrl: { type: string }
        skills:
          type: array
          items: { type: string }
    EmployerProfile:
      type: object
      required: [userId, companyName, website, industry, location, description, logoUrl, updatedAt]
      properties:
        userId: { type: string }
        companyName: { type: string }
        website: { type: string }
        industry: { type: string }
        location: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        updatedAt: { type: string, format: date-time }
    EmployerProfileUpdate:
      type: object
      properties:
        companyName: { type: string }
        website: { type: string }
        industry: { type: string }
        location: { type: string }
        description: { type: string }
        logoUrl: { type: string }
...