# lib/api-zod/src/generated/types/employerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployerProfile {
  userId: string;
  companyName: string;
  website: string;
  industry: string;
  location: string;
  description: string;
  logoUrl: string;
  updatedAt: Date;
}