# lib/api-zod/src/generated/types/seekerProfileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SeekerProfileUpdate {
  fullName?: string;
  headline?: string;
  location?: string;
  phone?: string;
  bio?: string;
  resumeUrl?: string;
  skills?: string[];
}